+++ src/Web/index.js
@@ -10,7 +10,7 @@ const client = new Client({
 	id: process.env.CLIENT_ID,
 	secret: process.env.CLIENT_SECRET,
 	redirectURI: 'https://tanaka.1chi.tk/auth/callback',
-	scopes: ['identify', 'guilds.join'],
+	scopes: ['identify'],
 });
 
 // eslint-disable-next-line valid-jsdoc
@@ -72,9 +72,6 @@ module.exports = (c) => {
 
 		const code = req.query.code.toString();
 		const key = await client.getAccess(code);
-		const user = await client.getUser(key);
-
-		await c.addGuildMember(key, user.id);
 
 		res.cookie('discordToken', key).redirect('/');
 	});